Move resources to the service packages (#8126)
